+++ bika/health/browser/doctors/getdoctors.py
@@ -5,14 +5,13 @@
 # Copyright 2018 by it's authors.
 # Some rights reserved. See LICENSE.rst, CONTRIBUTORS.rst.
 
-from Products.CMFCore.utils import getToolByName
-from bika.health import bikaMessageFactory as _
-from bika.health.permissions import *
-from bika.lims import bikaMessageFactory as _b
-from bika.lims.browser import BrowserView
-from operator import itemgetter
 import json
+from operator import itemgetter
+
 import plone
+from bika.lims import api
+from bika.lims.browser import BrowserView
+from bika.lims.interfaces import ILabContact, IClient
 
 
 class ajaxGetDoctors(BrowserView):
@@ -25,21 +24,27 @@ class ajaxGetDoctors(BrowserView):
         nr_rows = self.request['rows']
         sord = self.request['sord']
         sidx = self.request['sidx']
-
         rows = []
 
-        pc = self.portal_catalog
-        proxies = pc(portal_type="Doctor")
-        for doctor in proxies:
-            doctor = doctor.getObject()
-            if self.portal_workflow.getInfoFor(doctor, 'inactive_state', 'active') == 'inactive':
+        query = dict(portal_type="Doctor", inactive_state="active")
+        client = self.get_current_client()
+        if client:
+            # Search those Doctors that are assigned to the same client or
+            # that do not have any client assigned
+            query["getPrimaryReferrerUID"] = [api.get_uid(client), None]
+
+        doctors = api.search(query, 'portal_catalog')
+        for doctor in doctors:
+            doctor_id = doctor.id
+            doctor_title = doctor.Title
+            search_val = ('{} {}'.format(doctor_id, doctor_title)).lower()
+            if searchTerm not in search_val:
                 continue
-            if doctor.Title().lower().find(searchTerm) > -1 \
-            or doctor.getDoctorID().lower().find(searchTerm) > -1:
-                rows.append({'Title': doctor.Title() or '',
-                             'DoctorID': doctor.getDoctorID(),
-                             'DoctorSysID': doctor.id,
-                             'DoctorUID': doctor.UID()})
+
+            rows.append({'Title': doctor.Title() or '',
+                         'DoctorID': doctor.getDoctorID(),
+                         'DoctorSysID': doctor.id,
+                         'DoctorUID': doctor.UID()})
 
         rows = sorted(rows, cmp=lambda x, y: cmp(x.lower(), y.lower()), key = itemgetter(sidx and sidx or 'Title'))
         if sord == 'desc':
@@ -52,3 +57,21 @@ class ajaxGetDoctors(BrowserView):
                'rows': rows[(int(page) - 1) * int(nr_rows): int(page) * int(nr_rows)]}
 
         return json.dumps(ret)
+
+    def get_current_client(self, default=None):
+        """Returns the client the current user belongs to
+        """
+        user = api.get_current_user()
+        roles = user.getRoles()
+        if 'Client' not in roles:
+            return default
+
+        contact = api.get_user_contact(user)
+        if not contact or ILabContact.providedBy(contact):
+            return default
+
+        client = api.get_parent(contact)
+        if not client or not IClient.providedBy(client):
+            return default
+
+        return client